requirements.txt
pandas
requests
pillow
openpyxl
